# proxy: avoid copying bodies multiple times